Version 3.0 beta     Added a new option to the config file: 	blacklist: blocks a domain         hostfiles: a list that contains the source files
    The host file now have some comments
    The configuration was moved to lib.config module

+++ lib/config.py
@@ -1,46 +1,42 @@
 import json
-from os.path import exists
+import os
 
-def read_config(config_file, *keys):
-    data = None
+import lib.termcolor as termcolor
 
-    with open(config_file) as raw_config:
+
+__BASE_DIR = os.path.dirname(os.path.dirname(__file__))
+CONFIG = os.path.join(__BASE_DIR, 'config.json')
+DATABASE = os.path.join(__BASE_DIR, 'adaway.db')
+
+def read(key):
+    with open(CONFIG) as raw_config:
         json_file = json.load(raw_config)
 
-        if len(keys) > 1:
-            data = list()
+        return json_file[key]
 
-            for key in keys:
-                data.append(json_file[key])
-        else:
-            data = json_file[keys[0]]
+def write():
+    if os.path.exists(CONFIG):
+        return
 
-
-    return data
-
-def write_default_config(filename):
-    with open(filename, 'w') as config_file:
-        raw_config = \
-            {
-                "blacklist": [
-                    "http://adaway.org/hosts.txt",
-                    "http://hosts-file.net/ad_servers.asp",
-                    "http://pgl.yoyo.org/adservers/serverlist.php?hostformat=hosts&showintro=0&mimetype=plaintext",
-                    "http://winhelp2002.mvps.org/hosts.txt",
-                    "http://someonewhocares.org/hosts/hosts"
-                ],
-                "custom_hosts": {
-                    "localhost": "127.0.0.1",
-                    "ip6-localhost ip6-loopback": "::1",
-                    "ip6-localnet": "fe00::0",
-                    "ip6-mcastprefix": "ff00::0",
-                    "ip6-allnodes": "ff02::1",
-                    "ip6-allrouters": "ff02::2"
-                },
-                "whitelist": [
-                    "adf.ly",
-                    "www.linkbucks.com",
-                ]
-            }
+    with open(CONFIG, 'w') as config_file:
+        raw_config = {
+            'host_files': [
+                'http://adaway.org/hosts.txt',
+                #'http://hosts-file.net/ad_servers.asp',
+                #'http://pgl.yoyo.org/adservers/serverlist.php?hostformat=hosts&showintro=0&mimetype=plaintext',
+                #'http://winhelp2002.mvps.org/hosts.txt',
+                #'http://someonewhocares.org/hosts/hosts'
+            ],
+            'blacklist': [
+                'youtube.com'
+            ],
+            'custom_hosts': {
+                'saegusa': '128.0.0.1'
+            },
+            'whitelist': [
+                'adf.ly',
+                'www.linkbucks.com'
+            ]
+        }
 
         json.dump(raw_config, config_file, indent=4)
